+++ src/primitive/int_float.h
@@ -22,7 +22,7 @@ void IntFloat_free(void *ptr);
 int64_t IntFloat_get(void *ptr);
 void IntFloat_set(void *ptr, int64_t value);
 bool IntFloat_compareAndSet(void *ptr, int64_t expected, int64_t value);
-uint32_t IntFloat_type(void *ptr);
+uint64_t IntFloat_type(void *ptr);
 size_t IntFloat_length(void *ptr);
