+++ src/struct/collection.c
@@ -16,7 +16,7 @@
  * STRUCT FIELDS (Mirroring struct/collection.h):
  * ----------------------------------------------------------------------------
  *   Collection {
- *     uint32_t typeId; // mirror of the block-header type (for debug)
+ *     uint64_t typeId; // mirror of the block-header type (for debug)
  *     uint32_t activeCount; // number of live elements
  *     uint32_t elementClass; // class of elements (Map: key class)
  *     uint32_t stride; // bytes per element (Map: val class)
@@ -69,7 +69,7 @@ static void writeSlotAt(uint8_t *data, size_t stride, uint64_t value) {
     }
 }
 
-uint32_t Collection_type(Collection *c) {
+uint64_t Collection_type(Collection *c) {
     if (!c) return 0;
     return Memory_type(c);
 }
